raft.go: simplify checking for majority of votes
I was looking how Raft would deal with split-votes. A candidate that
receives votes from a majority of the servers in the full cluster
becomes the leader. But what means majority? Obviously it means more
than half of the servers, which is mathematically `majority >=
floor(n/2) + 1` (where `n` is the number of servers in the full
cluster).

I thought rewriting the if clause that represents the mathetmical
notation would make it easier for the reader to understand how we check
for the majority of the votes. Because integer divisions in Go already
perform floor division, we only need to make a small change.

All the tests are passing this change, so it should be good to merge:

```zsh
$ for x (part1 part2 part3); do echo "Testing ${x} ..." && pushd $x && go test ./... && popd; done
Testing part1 ...
ok      github.com/eliben/raft  8.656s
Testing part2 ...
ok      github.com/eliben/raft  15.168s
Testing part3 ...
ok      github.com/eliben/raft  26.035s
```